# swagger.yaml
openapi: 3.0.0
paths:
...
  /signin:
    post:
      operationId: AuthController_signin
      summary: Войти в систему
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInRequestBody'
      responses:
        '201':
          description: Успешный вход в систему
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninResponseBodyOK'
        '401':
          description: Неверное имя пользователя или пароль
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninResponseBodyNotOK'
      tags: &ref_0
        - auth
  /signup:
    post:
      operationId: AuthController_signup
      summary: Регистрация
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequestBody'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninResponseBodyOK'
        '400':
          description: Некорректные данные
        '409':
          description: Аккаунт уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponseBodyNotOK'
      tags: *ref_0
  /refresh-token:
    post:
      operationId: AuthController_refreshToken
      summary: Обновить токен
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequestBody'
      responses:
        '201':
          description: accessToken успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refreshTokenResponseBodyOK'
        '401':
          description: Невалидный refreshToken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/refreshTokenResponseBodyNotOK'
      tags: *ref_0
  /yandex/exchange:
    post:
      operationId: AuthController_exchangeCodeYandex
      summary: Авторизация через Yandex
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeFlowAuthRequestBody'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninResponseBodyOK'
      tags: *ref_0
  /mailru/exchange:
    post:
      operationId: AuthController_exchangeCodeMail
      summary: Авторизация через Mail
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CodeFlowAuthRequestBody'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigninResponseBodyOK'
      tags: *ref_0
...
  /reset-password:
    post:
      operationId: AuthController_resetPassword
      summary: Сброс пароля
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestBody'
      responses:
        '201':
          description: Ссылка для сброса пароля отправлена на указанный Email
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResponseBodyOK'
        '400':
          description: Некорректные данные
        '404':
          description: Пользователь с указанным Email не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetPasswordResponseBodyNotFound'
      tags: *ref_0
...
  /accounts:
    get:
      operationId: AccountController_findAll
      summary: Получить все аккаунты
      parameters: []
      responses:
        '200':
          description: Запрос выполнен успешно
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SingleAccountResponseBodyOK'
        '403':
          description: Отказ в доступе
      tags: &ref_6
        - accounts
      security: &ref_7
        - bearer: []
  /accounts/{id}:
    patch:
      operationId: AccountController_update
      summary: Обновить данные аккаунта по id
      parameters:
        - name: id
          required: true
          in: path
          description: Индификатор аккаунта
          example: 64f81ba37571bfaac18a857f
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountUpdateRequestBody'
      responses:
        '200':
          description: Аккаунт успешно обновлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SingleAccountResponseBodyOK'
        '400':
          description: Неверный запрос
        '403':
          description: Отказ в доступе
        '404':
          description: Ресурс не найден
      tags: *ref_6
      security: *ref_7
...
info:
  title: API BotKits
  description: Ручки для команды frontend
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    SignInRequestBody:
      type: object
      properties:
        email:
          type: string
          example: test@mail.ru
          description: Почта пользователя
        password:
          type: string
          example: '123'
          description: Пароль пользователя
      required:
        - email
        - password
    SigninResponseBodyOK:
      type: object
      properties:
        _id:
          type: string
          example: 650b396ed4201e5ca499f3b5
        type:
          type: string
          example: local
        role:
          type: string
          example: user
        credentials:
          type: object
          example:
            email: test@mail.ru
            accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
            refreshToken: eyJhbGciOiJIUzI1NiIsInR5cPCI6IkpXVCJ9
        profile:
          type: object
          example:
            _id: 650b396dd4201e5ca499f3b3
            username: test
            phone: '+79999999999'
            avatar: https://i.pravatar.cc/300
            balance: 0
        success:
          type: boolean
          example: true
      required:
        - credentials
        - profile
    SigninResponseBodyNotOK:
      type: object
      properties:
        message:
          type: string
          example: Неверное имя пользователя или пароль
          description: Сообщение об ошибке
        error:
          type: string
          example: Unauthorized
          description: Тип ошибки
        statusCode:
          type: number
          example: 401
          description: HTTP-статус код
    SignUpRequestBody:
      type: object
      properties:
        password:
          type: string
          example: '123'
          description: Пароль пользователя
        email:
          type: string
          example: test@mail.ru
          description: Email пользователя
        phone:
          type: string
          example: '+79999999999'
          description: Телефонный номер пользователя
        username:
          type: string
          example: test
          description: Имя пользователя
      required:
        - password
        - email
        - phone
        - username
    SignupResponseBodyNotOK:
      type: object
      properties:
        message:
          type: string
          example: Аккаунт уже существует
          description: Сообщение об ошибке
        error:
          type: string
          example: Conflict
          description: Тип ошибки
        statusCode:
          type: number
          example: 409
          description: HTTP-статус код
    RefreshRequestBody:
      type: object
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
          description: Выданный рефреш токен
      required:
        - refreshToken
    refreshTokenResponseBodyOK:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
          description: accessToken по умолчанию действует 1 день
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
          description: refreshToken по умолчанию действует 7 день
    refreshTokenResponseBodyNotOK:
      type: object
      properties:
        message:
          type: string
          example: Невалидный refreshToken
          description: Сообщение об ошибке
        error:
          type: string
          example: Unauthorized
          description: Тип ошибки
        statusCode:
          type: number
          example: 401
          description: HTTP-статус код
    CodeFlowAuthRequestBody:
      type: object
      properties:
        codeAuth:
          type: string
          example: '432532'
          description: Код авторизации yandex
      required:
        - codeAuth
...
    ResetPasswordRequestBody:
      type: object
      properties:
        email:
          type: string
          example: test@mail.ru
          description: Почта пользователя
      required:
        - email
    ResetPasswordResponseBodyOK:
      type: object
      properties:
        message:
          type: string
          example: 'Ссылка на сброс пароля отправлена на ваш email: test@mail.ru'
          description: Сообщение об успешном сбросе пароля
    ResetPasswordResponseBodyNotFound:
      type: object
      properties:
        message:
          type: string
          example: Пользователь с указанным Email не найден
          description: Сообщение об ошибке
        error:
          type: string
          example: Not Found
          description: Тип ошибки
        statusCode:
          type: number
          example: 404
          description: HTTP-статус код
...
    SingleAccountResponseBodyOK:
      type: object
      properties:
        _id:
          type: string
          example: 6516ab8f2c612e8cce53e45f
        type:
          type: string
          example: local
        role:
          type: string
          example: user
        credentials:
          type: object
          example:
            email: test@mail.ru
            accessToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
            refreshToken: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9
        profile:
          type: string
          example: 6516ab8f2c612e8cce53e45d
        success:
          type: boolean
          example: true
      required:
        - credentials
...
    AccountUpdateRequestBody:
      type: object
      properties:
        type:
          type: string
          example: local
        role:
          type: string
          example: user
        credentials:
          type: object
          example:
            email: test@mail.ru
      required:
        - credentials
...